# Remove []byte copy in framer