… testdata/frames/bench_parse_result.gz