Dont defensively copy framers buf when reading bytes during parseFrame.

Updates #492