@Zariel 2015-10-17
Before and after

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkParseRowsFrame-4     76868         34321         -55.35%

benchmark                     old allocs     new allocs     delta
BenchmarkParseRowsFrame-4     609            110            -81.94%

benchmark                     old bytes     new bytes     delta
BenchmarkParseRowsFrame-4     31904         20304         -36.36%
```